+++ app/js/history.js
@@ -182,7 +182,7 @@ const handleRequest = (type, start, end) => {
     } else if (type === "all") {
         clearHistory()
     }
-    TABS.currentPage().getWebContents().send("history-list", history.map(h => {
+    TABS.webContents(TABS.currentPage()).send("history-list", history.map(h => {
         h.icon = FAVICONS.forSite(h.url)
         return h
     }))